Håndtere dager per uke med desimaler (#553)
* Håndtere dager per uke med desimaler

* Fjernet muligheten for å endre til desimaler fra frontend

+++ core/src/main/kotlin/no/nav/amt/tiltak/core/domain/tiltak/Deltaker.kt
@@ -20,7 +20,7 @@ data class Deltaker(
 	val sluttDato: LocalDate?,
 	val status: DeltakerStatus,
 	val registrertDato: LocalDateTime,
-	val dagerPerUke: Int? = null,
+	val dagerPerUke: Float? = null,
 	val prosentStilling: Float? = null,
 	val innsokBegrunnelse: String? = null,
 	val endretDato: LocalDateTime
